Adding more bits
